requirements.txt
Pillow
python-multipart
torch
transformers
fastapi
uvicorn
python-keycloak
python-jose[cryptography]